+++ application/views.py
@@ -1,8 +1,8 @@
-from myapplication import application
-from myapplication.database import db
+from application import application
+from application.database import db
 from flask import render_template, request
-from myapplication.models import Data
-from myapplication.forms import EnterDBInfo, RetrieveDBInfo
+from application.models import Data
+from application.forms import EnterDBInfo, RetrieveDBInfo
 from environs import Env
 
 env = Env()
