+++ test.py
@@ -1,4 +1,4 @@
-from myapplication import application
+from application import application
 with application.test_client() as c:
     response = c.get('/')
     assert response.status_code == 200
